Kernel: Move Partition code files to the Storage folder
This folder is more appropriate for these files.
